fix(agent): fix critical tool call ID mismatch and message format issues
- ToolCallDelta: change from id-based to index-based matching to fix
  Claude SSE stream argument accumulation (Critical #1)
- Add MessageContent::ToolUse variant so assistant messages include
  tool_use blocks required by Claude/OpenAI APIs (Critical #2)
- Fix OpenAI provider: tool results now emit as role:tool with
  tool_call_id instead of role:user (Important #3)
- Fix OpenAI content serialization from array to string + tool_calls
  array format (Important #4)
- Add TerminalGuard RAII to restore terminal on panic (Important #6)
- Fix scroll_offset from u16 to usize (Minor #8)
- Fix truncate() no longer assumes first message is system (Minor #7)

+++ crates/cli/src/tui/layout.rs
@@ -91,7 +91,7 @@ fn draw_chat_panel(f: &mut Frame, app: &App, area: Rect) {
     let paragraph = Paragraph::new(lines)
         .block(Block::default().borders(Borders::ALL).title(" Chat "))
         .wrap(Wrap { trim: false })
-        .scroll((app.scroll_offset, 0));
+        .scroll((app.scroll_offset as u16, 0));
 
     f.render_widget(paragraph, area);
 }
